requirements.txt
cloudpickle==0.4.0
exifread==2.1.2
gpxpy==1.1.2
loky==1.2.1
networkx==1.11
numpy
pyproj==1.9.5.1
pytest==3.0.7
python-dateutil==2.6.0
PyYAML==3.12
repoze.lru==0.7
scipy
six
xmltodict==0.10.2
nudged=0.3.1
pandas=0.23.4
opencv-python=3.4.3.18